Added pytest fixture that sets seeds before tests
This fixture will automatically set the python, numpy and torch seeds
before running each test in the hope of being able to get some
reprodicibility.  There are certain module in torch that may mean the
tests are not 100% identical each time, but for the most part this
should already help with randomly failing tests.